sr(api): move handle-training-recommendation in devcomp
