+++ api/src/devcomp/domain/usecases/handle-training-recommendation.js
